Cursos/pythonProject/requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.2
pydantic==2.6.1